feat(seo): add generic page seo composable
